requirements.txt
# IDVestAI — Python Dependencies
#
# Install all at once:
#   pip install -r requirements.txt
#
# GPU (NVIDIA) users — after installing requirements, run:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121

# ── Core AI ──────────────────────────────────────────────────────────────────
ultralytics>=8.2.0          # YOLOv8 — model training and inference
torch>=2.1.0                # PyTorch backend (CPU version; see note above for GPU)
torchvision>=0.16.0

# ── Computer Vision ──────────────────────────────────────────────────────────
opencv-python-headless>=4.9.0 # Real-time video capture and image processing
Pillow>=10.0.0              # Image I/O (used internally by Ultralytics)

# ── Dashboard ────────────────────────────────────────────────────────────────
flask>=3.0.0                # WSGI web framework for live MJPEG streaming
streamlit>=1.35.0           # Interactive dashboard UI (streamlit run streamlit_app.py)

# ── FastAPI backend (optional) ───────────────────────────────────────────────
fastapi>=0.111.0            # ASGI web framework
uvicorn[standard]>=0.29.0   # High-performance ASGI server
pydantic>=2.7.0             # Request / response data validation
python-multipart>=0.0.9     # Form / file upload support

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0        # Load .env file for environment variables
numpy>=1.26.0               # Numerical arrays (used by OpenCV and PyTorch)

# ── Development / testing ─────────────────────────────────────────────────────
pytest>=8.0.0               # Unit testing framework
httpx>=0.27.0               # Async HTTP client (for FastAPI TestClient)
